refactor(bindings): type trade status as a shared TradeStatus enum
