+++ packages/bindings/src/shared.ts
@@ -27,6 +27,19 @@ export enum OrderType {
   FAK = 'FAK',
 }
 
+/**
+ * Lifecycle status of a trade, from creation through on-chain settlement.
+ * `Confirmed` and `Failed` are the terminal states.
+ */
+export enum TradeStatus {
+  Matched = 'TRADE_STATUS_MATCHED',
+  MatchedNotBroadcasted = 'TRADE_STATUS_MATCHED_NOT_BROADCASTED',
+  Mined = 'TRADE_STATUS_MINED',
+  Confirmed = 'TRADE_STATUS_CONFIRMED',
+  Retrying = 'TRADE_STATUS_RETRYING',
+  Failed = 'TRADE_STATUS_FAILED',
+}
+
 function toTaggedString<T extends string>(value: string): T {
   return value as T;
 }
@@ -394,6 +407,20 @@ export const NotificationIdSchema = z
 export const CommentParentEntityTypeSchema = z.enum(CommentParentEntityType);
 export const OrderSideSchema = z.enum(OrderSide);
 export const OrderTypeSchema = z.enum(OrderType);
+// Trade statuses arrive in two wire forms: REST endpoints serialize the raw
+// prefixed constants ("TRADE_STATUS_CONFIRMED") while the user websocket
+// channel serializes plain values ("CONFIRMED"). Normalize both to the enum.
+export const TradeStatusSchema = z.preprocess((value) => {
+  if (typeof value !== 'string' || value.startsWith('TRADE_STATUS_')) {
+    return value;
+  }
+
+  const normalized = Object.values(TradeStatus).find(
+    (status) => status.slice('TRADE_STATUS_'.length) === value,
+  );
+
+  return normalized ?? value;
+}, z.enum(TradeStatus));
 export const PaginationCursorSchema = z.custom<PaginationCursor>(
   (value) => typeof value === 'string' && value.length > 0,
   'Expected a non-empty pagination cursor',
